# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API for managing contacts information",
    "version": "1.0.0"
  },
  "host": "cse-341spring.onrender.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "Endpoints for managing contacts"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Paul"
        },
        "lastName": {
          "type": "string",
          "example": "Lukwago"
        },
        "email": {
          "type": "string",
          "example": "paullukwago@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "blue"
        },
        "birthday": {
          "type": "string",
          "example": "1999-01-01"
        }
      }
    },
    "ContactInput": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Daniel"
        },
        "email": {
          "type": "string",
          "example": "johndaniel@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "blue"
        },
        "birthday": {
          "type": "string",
          "example": "1995-08-01"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "favoriteColor",
        "birthday"
      ]
    },
    "ContactResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "682274ddd0a5b29bade88762"
        },
        "firstName": {
          "type": "string",
          "example": "Edward"
        },
        "lastName": {
          "type": "string",
          "example": "Yawe"
        },
        "email": {
          "type": "string",
          "example": "edwardyawe@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "blue"
        },
        "birthday": {
          "type": "string",
          "example": "2000-01-21"
        }
      }
    },
    "NewContactResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "682274ddd0a5b29bade88762"
        },
        "message": {
          "type": "string",
          "example": "Contact created successfully"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Error message"
        },
        "message": {
          "type": "string",
          "example": "Detailed error information"
        }
      }
    }
  }
}